fix: bluetooth advertisement is working now
When comparing BLEA raw data, noticed that the one from core_lib was
different than a legit android device (mine used an Advertising
Extension) which seems to cause an issue. Fix is to remove the
service_uuid from the advertisement. Power and oneshot are not needed so
removing as they are of no use.

Signed-off-by: Martichou <m@rtin.fyi>