feat(BREAKING): remove legacy variant
Due to Github removing support for the Ubuntu-20.04 os image,
I'm removing the legacy variant that was kept alive only for
backward compatibility. RQuickShare will sadly align with Tauri
v2 system requirements.

The previous release v0.11.5 will be the last one to include
the legacy binary. This version is still working and will be the
one people wanting to use RQuickShare on "older" system shall
default to.

Signed-off-by: Martichou <m@rtin.fyi>