Add support for custom device names
Adds an option in the settings menu for the user to specify a custom device name.
Empty or entirely whitespace values will be treated as unset.

On the ts/vue side, names are stored as string/undefined with normalization of empty/whitespace strings to undefined.
On the rust side, names are stored as Option<String>.

If a valid custom device name is specified, it will be used in place of the hostname in mdns advertisements.
If no device name or an invalid device name is specified, the hostname will be used as a fallback.

Fixes #310.